requirements.txt
PyGithub
packaging~=21.0
requests
